fix(agents): keep child_process out of the prompt modules (AGT-4088)
Rendering the sibling block made the prompt templates import siblingWork,
which calls `promisify(execFile)` at module scope. Any suite that partially
mocks child_process — scheduler, prProcessor — then failed to load at all,
with "No 'execFile' export is defined on the child_process mock". CI caught
it; two scheduler suites collected zero tests.

Splits the pure half (parsing, selection, rendering) into siblingWorkFormat
and leaves the collector — and canonicalPath, which does its own I/O — in
siblingWork, which re-exports the pure names so the collector's API is
unchanged. The prompts now import only the pure module, which is the right
dependency anyway: a template has no business reaching for a subprocess.

Adds a test that mocks child_process empty and imports both locales, which is
the failure CI saw. Reintroducing the old import makes it fail.
